refactor: split harn-vm value module
